# libs/domains/agenstra/backend/feature-agent-controller/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Controller API
  version: 1.0.0
  description: HTTP API for managing clients and proxied agent operations
  contact:
    name: ForePath
    email: hi@forepath.io
    url: https://forepath.io
servers:
  - url: http://localhost:3100/api
    description: Local development
  - url: http://localhost:3100
    description: Root (OpenTelemetry metrics; outside /api prefix)
security:
  - bearerAuth: []
paths:
...
  /clients/provisioning/providers/{providerType}/server-types:
    get:
      summary: Get available server types for a provisioning provider
      operationId: getServerTypes
      parameters:
        - in: path
          name: providerType
          required: true
          schema:
            type: string
            enum: [hetzner, digital-ocean]
          description: The provisioning provider type (e.g., 'hetzner', 'digital-ocean')
      responses:
        '200':
          description: Array of server types with specifications and pricing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServerType'
        '400':
          description: Provider type not found or invalid
  /clients/provisioning/providers/{providerType}/locations:
    get:
      summary: Get available geography options for a provisioning provider
      operationId: getProviderLocations
      parameters:
        - in: path
          name: providerType
          required: true
          schema:
            type: string
            enum: [hetzner, digital-ocean]
          description: The provisioning provider type (e.g., 'hetzner', 'digital-ocean')
      responses:
        '200':
          description: Array of locations/regions with human-readable labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProviderLocation'
        '400':
          description: Provider type not found or invalid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
...
  schemas:
...
    ProviderLocation:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          description: Technical slug used for provisioning (e.g. fsn1, fra1)
        name:
          type: string
          description: Human-readable display name from provider API or static fallback
        city:
          type: string
          description: City when provided by the provider API (Hetzner)
        country:
          type: string
          description: ISO country code when provided by the provider API
    ServerType:
      type: object
      required: [id, name, cores, memory, disk]
      properties:
        id:
          type: string
          description: Unique identifier for the server type (e.g., 'cx11', 'cpx11')
        name:
          type: string
          description: Human-readable name (e.g., 'CX11', 'CPX11')
        cores:
          type: integer
          description: Number of CPU cores
        memory:
          type: number
          description: Amount of RAM in GB
        disk:
          type: integer
          description: Disk size in GB
        priceMonthly:
          type: number
          description: Price per month in EUR (or provider's currency)
        priceHourly:
          type: number
          description: Price per hour in EUR (or provider's currency)
        description:
          type: string
          description: Additional description or notes
...